+++ src/Applet/appletCommand.ts
@@ -2,6 +2,7 @@ import ICommand from "../Command/ICommand";
 import { appletGenerate } from './Generate/appletGenerateCommand';
 import { appletUpload } from './Upload/appletUploadCommand';
 import { appletStart } from './Start/appletStartCommand';
+import { appletTest } from "./Test/appletTestCommand";
 
 export const applet: ICommand = {
 	name: 'applet',
@@ -11,6 +12,7 @@ export const applet: ICommand = {
 		appletGenerate,
 		appletUpload,
 		appletStart,
+		appletTest,
 	],
 	async run() {
 		throw new Error('Unknown command');